umplementing interfaces

+++ Interfaces/IWindPoweredOptions.cs
@@ -4,6 +4,6 @@ public interface IWindPoweredOptions
 
 {
     public int Production { get; set; }
-    public int MaximumWind { get; set; }
+    public float MaximumWind { get; set; }
 
 }
